Added action to API calls

+++ src/Rest/BaseApiHandler.php
@@ -13,37 +13,37 @@ abstract class BaseApiHandler
   public function __construct($slug = NULL)
   {
     $this->slug = $slug;
-    $this->defaultHeaders = array();
+    $this->defaultHeaders = [];
   }
 
   public function get(Request $request)
   {
-    return new Response(null, 405, array());
+    return new Response(null, 405, []);
   }
 
   public function post(Request $request)
   {
-    return new Response(null, 405, array());
+    return new Response(null, 405, []);
   }
 
   public function put(Request $request)
   {
-    return new Response(null, 405, array());
+    return new Response(null, 405, []);
   }
 
   public function patch(Request $request)
   {
-    return new Response(null, 405, array());
+    return new Response(null, 405, []);
   }
 
   public function delete(Request $request)
   {
-    return new Response(null, 405, array());
+    return new Response(null, 405, []);
   }
 
   public function options(Request $request)
   {
-    return new Response(null, 405, array());
+    return new Response(null, 405, []);
   }
 
   private final function setDefaultHeaders(Response $response)
@@ -54,39 +54,56 @@ abstract class BaseApiHandler
     }
   }
 
-  public final function handle(Request $request)
+  public final function handle(Request $request, $action = NULL)
   {
     $response;
     $method = $request->getMethod();
 
-    if($method === 'GET')
+    // strip all none alpha numeric characters, an action is a function name
+    $action = empty($action) ? '' : preg_replace('/[^A-Za-z0-9 ]/', '', $action);
+
+    if($method === 'GET' && empty($action))
     {
       $response = $this->get($request);
     }
     else if($method === 'POST')
     {
-      $response = $this->post($request);
+      if(empty($action))
+      {
+        // Normal post, lets handle it via the POST method
+        $response = $this->post($request);
+      }
+      else
+      {
+        $action = ucfirst(strtolower($action));
+        $method = 'action'.$action;
+
+        if(is_callable([$this, $method]))
+        {
+          $response = $this->$method($request);
+        }
+      }
     }
-    else if($method === 'PUT')
+    else if($method === 'PUT' && empty($action))
     {
       $response = $this->put($request);
     }
-    else if($method === 'PATCH')
+    else if($method === 'PATCH' && empty($action))
     {
       $response = $this->patch($request);
     }
-    else if($method === 'DELETE')
+    else if($method === 'DELETE' && empty($action))
     {
       $response = $this->delete($request);
     }
-    else if($method === 'OPTIONS')
+    else if($method === 'OPTIONS' && empty($action))
     {
       $response = $this->options($request);
     }
 
     if(empty($response))
     {
-      $response = new Response(null, 400, array());
+      $response = new Response(null, 400, []);
     }
     else
     {
